+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - replaced MD5 hashing of `fsm.StateMachine` with blake2b.
+- improved readability of `ValidationError` messages in `fsm.StateMachine`.
 
 ## [0.1.0a8] - 2026-03-04